fix: decouple buckets from app code
It is a good practice to avoid defining buckets with values that may
change. By hardcoding bucket definitions this way, we avoid breaking
grafana boards when someone decides to adjust a timeout.

If we ever need to change buckets, or add new one with bigger max,
but keep old ones intact, we can define them by hand.